requirements.txt
pymongo==3.6.0
python_telegram_bot==9.0.0
coinmarketcap==4.2.1
pytz==2017.3
timezonefinder==2.1.2
